Allow containers to be canceled and restarted
We don't have to use the docker API in a blocking way. We do have to
implement our own "wait" command, since the API is only interested in
waiting for a container to exit. This is much cleaner than using the
contexts and cancellations.

We also move the docker host into its own "actor" which can be passed
messages. This pattern has worked well in Rust, and so will likely work
well here.