Support context for run function
This is used to cancel the run function, but not the docker arguments. I
originally tried passing the `run` context to the docker commands, but
of course when the context was cancelled, then the docker commands
wouldn't run because their context had expired.

I still have some learning to do with contexts in Go.